Enable multi-site models
Reworked the vacancy site tracking to be simpler and require less user
input. Also enabled the ability to have multiple adsorption sites in the
model. Did some bug squashing and added a few comments (more coming).
Still to do is rework the vacancy site area so that the site area is a
property of the vacancy, rather than the model. This would eliminate the
need of the Model to pass Asite to the Reaction objects.